+++ tee-worker/ts-tests/integration-tests/di_vc.test.ts
@@ -9,7 +9,7 @@ import {
     sendRequestFromTrustedCall,
     createSignedTrustedCallRequestVc,
 } from './common/di-utils'; // @fixme move to a better place
-import type { IntegrationTestContext } from './common/common-definitions';
+import type { IntegrationTestContext } from './common/common-types';
 import { aesKey } from './common/call';
 import { LitentryPrimitivesIdentity } from 'sidechain-api';
 import { subscribeToEventsWithExtHash } from './common/transactions';
